Add option for dropping unused factors from legend
In the original behavior, by default, unused factor levels are removed
from the legend. This commit adds a logical option to toggle this
behavior, with the default being the original behavior to drop levels.
Some lines of code are also changed to use toptable$Sig rather than
factor(Sig), as the original variable in my testing was empty and did
not produce the desired result with the new option (though did not seem
to produce bugs otherwise).